backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
requests
entsoe-py
pulp
pydantic
python-dotenv
pytest
pytest-mock
httpx